pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "process-scanned-photos"
version = "0.1.0"
description = "A tool for processing scanned photos"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    { name = "Travis Whitfield" }
]
dependencies = [
    "pillow>=10.0",
    "numpy>=1.24",
    # Add other runtime dependencies here
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
